Refactor unit tests to a self-contained directory
