+++ src/swoole/phpunit.xml.dist
@@ -11,8 +11,8 @@
     <ini name="error_reporting" value="-1"/>
   </php>
   <testsuites>
-    <testsuite name="Test Suite">
-      <directory>./tests</directory>
+    <testsuite name="Unit">
+      <directory>./tests/Unit</directory>
     </testsuite>
   </testsuites>
   <filter>
