+++ src/swoole/tests/Unit/LaravelRunnerTest.php
@@ -2,22 +2,22 @@
 
 declare(strict_types=1);
 
-namespace Runtime\Swoole\Tests;
+namespace Runtime\Swoole\Tests\Unit;
 
+use Illuminate\Contracts\Http\Kernel;
 use PHPUnit\Framework\TestCase;
+use Runtime\Swoole\LaravelRunner;
 use Runtime\Swoole\ServerFactory;
-use Runtime\Swoole\SymfonyRunner;
 use Swoole\Http\Request;
 use Swoole\Http\Response;
 use Swoole\Http\Server;
 use Symfony\Component\HttpFoundation\HeaderBag;
-use Symfony\Component\HttpKernel\HttpKernelInterface;
 
-class SymfonyRunnerTest extends TestCase
+class LaravelRunnerTest extends TestCase
 {
     public function testRun(): void
     {
-        $application = $this->createMock(HttpKernelInterface::class);
+        $application = $this->createMock(Kernel::class);
 
         $server = $this->createMock(Server::class);
         $server->expects(self::once())->method('start');
@@ -25,7 +25,7 @@ class SymfonyRunnerTest extends TestCase
         $factory = $this->createMock(ServerFactory::class);
         $factory->expects(self::once())->method('createServer')->willReturn($server);
 
-        $runner = new SymfonyRunner($factory, $application);
+        $runner = new LaravelRunner($factory, $application);
 
         self::assertSame(0, $runner->run());
     }
@@ -36,7 +36,7 @@ class SymfonyRunnerTest extends TestCase
         $sfResponse->headers = new HeaderBag(['X-Test' => 'Swoole-Runtime']);
         $sfResponse->expects(self::once())->method('getContent')->willReturn('Test');
 
-        $application = $this->createMock(HttpKernelInterface::class);
+        $application = $this->createMock(Kernel::class);
         $application->expects(self::once())->method('handle')->willReturn($sfResponse);
 
         $request = $this->createMock(Request::class);
@@ -48,7 +48,7 @@ class SymfonyRunnerTest extends TestCase
 
         $factory = $this->createMock(ServerFactory::class);
 
-        $runner = new SymfonyRunner($factory, $application);
+        $runner = new LaravelRunner($factory, $application);
         $runner->handle($request, $response);
     }
 }
